requirements.txt
fastapi
uvicorn[standard]
pyserial
fastapi_mail
python-dotenv